TwilioVoiceToAsr/requirements.txt
aiohttp==3.7.4
google-cloud-speech==2.17.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
